test(runtime): split performance gate policy coverage
